feat: implement OpenRGB integration with device management and effect controls
